# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quest API
  description: |
    REST API for Quest - a Jira-like project and task management system.

    ## Authentication
    All endpoints require authentication via session cookie (NextAuth.js) or Bearer token.

    ## Base URL
    - Production: `https://quest-manas-pradhans-projects.vercel.app`
    - Development: `http://localhost:3000`

    ## Notes for Mobile Developers
    This API specification documents the full capabilities of Quest. All endpoints are
    implemented REST API routes that wrap Next.js Server Actions, making them consumable
    by iOS, Android, and other platforms via standard HTTP requests.
  version: 1.0.0
  contact:
    name: Quest Development Team
servers:
  - url: https://quest-manas-pradhans-projects.vercel.app
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/workspaces:
    get:
      tags: [Workspaces]
      summary: List user's workspaces
      operationId: listWorkspaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Workspaces]
      summary: Create a new workspace
      operationId: createWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '400':
          description: Validation error or slug taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/workspaces/{workspaceId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'

    get:
      tags: [Workspaces]
      summary: Get workspace details
      operationId: getWorkspace
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceWithMembers'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Workspaces]
      summary: Update workspace
      operationId: updateWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceRequest'
      responses:
        '200':
          description: Workspace updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Workspaces]
      summary: Delete workspace
      operationId: deleteWorkspace
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Workspace deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only owner can delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================
  # Projects
  # ===========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT or session token
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
      description: Workspace ID (CUID)
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
    Forbidden:
      description: Not authorized to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "You don't have permission to perform this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not found"
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
        image:
          type: string
          nullable: true

    # ===========================================
    # Workspace Schemas
    # ===========================================
    Workspace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        image:
          type: string
          nullable: true
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkspaceWithMembers:
      allOf:
        - $ref: '#/components/schemas/Workspace'
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: '#/components/schemas/WorkspaceMember'
            owner:
              $ref: '#/components/schemas/UserSummary'
    WorkspaceMember:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        workspaceId:
          type: string
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        createdAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserSummary'
    WorkspaceRole:
      type: string
      enum: [ADMIN, DEVELOPER, TESTER, GUEST]
    CreateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: Auto-generated from name if not provided
    UpdateWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
    WorkspaceResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Workspace'

    # ===========================================
    # Project Schemas
    # ===========================================
...